Tests as a module, with Travis command.

+++ tests/sql_tests.py
@@ -4,11 +4,16 @@ import sys
 import unittest
 import warnings
 
-sys.path.insert(0, "../src")
+sys.path.insert(0, "src")
 
 from cs50.sql import SQL
 
+
 class SQLTests(unittest.TestCase):
+
+    def setUp(self):
+        if self.__class__.__name__ == "SQLTests":
+            self.skipTest("This is a base class; no tests.")
 
     def test_multiple_statements(self):
         self.assertRaises(RuntimeError, self.db.execute, "INSERT INTO cs50(val) VALUES('baz'); INSERT INTO cs50(val) VALUES('qux')")
@@ -164,12 +169,15 @@ class SQLTests(unittest.TestCase):
         self.assertEqual(self.db.execute("SELECT val FROM cs50"), [])
 
     def tearDown(self):
-        self.db.execute("DROP TABLE cs50")
+        self.db.execute("DROP TABLE IF EXISTS cs50")
         self.db.execute("DROP TABLE IF EXISTS foo")
         self.db.execute("DROP TABLE IF EXISTS bar")
 
     @classmethod
     def tearDownClass(self):
+        if not hasattr(self, "db"):
+            return
+
         try:
             self.db.execute("DROP TABLE IF EXISTS cs50")
         except Warning as e:
@@ -209,7 +217,7 @@ class SQLiteTests(SQLTests):
         print("\nSQLite tests")
 
     def setUp(self):
-        self.db.execute("CREATE TABLE cs50(id INTEGER PRIMARY KEY, val TEXT, bin BLOB)")
+        self.db.execute("CREATE TABLE IF NOT EXISTS cs50(id INTEGER PRIMARY KEY, val TEXT, bin BLOB)")
 
     @classmethod
     @contextlib.contextmanager
@@ -229,7 +237,7 @@ class SQLiteTests(SQLTests):
 
     def test_parameter_warnings_deletes(self):
         logging.getLogger("cs50").disabled = False
-        
+
         deletes = [
             ["?", "", ["test"]],
             ["test", "WHERE ?=3", ["id"]]
@@ -245,7 +253,7 @@ class SQLiteTests(SQLTests):
                 except RuntimeError:
                     pass
 
-            self.assertTrue(True in map(lambda o: "This may cause errors" in o, cm.output))
+                self.assertTrue(True in map(lambda o: "This may cause errors" in o, cm.output))
 
         logging.getLogger("cs50").disabled = True
 
@@ -359,7 +367,7 @@ class SQLiteTests(SQLTests):
         for sel in selects:
             with self.assertLogs("cs50", level="DEBUG") as cm:
                 command = f"SELECT {sel[0]} FROM {sel[1]} {sel[2]}"
-                
+                    
                 try:
                     self.db.execute(command, *sel[3])
                 except RuntimeError:
@@ -571,9 +579,9 @@ class SQLiteTests(SQLTests):
 
 if __name__ == "__main__":
     suite = unittest.TestSuite([
-        unittest.TestLoader().loadTestsFromTestCase(SQLiteTests),
-        unittest.TestLoader().loadTestsFromTestCase(MySQLTests),
-        unittest.TestLoader().loadTestsFromTestCase(PostgresTests)
+        unittest.TestLoader().loadTestsFromTestCase(SQLiteTests)
+        #unittest.TestLoader().loadTestsFromTestCase(MySQLTests),
+        #unittest.TestLoader().loadTestsFromTestCase(PostgresTests)
     ])
 
     logging.getLogger("cs50").disabled = True
